// SGUEES-SPA/src/app/pages/SelectionHiring/sc-requisicion-observadores/models/sc-requisicion-observadores.ts
export interface ScRequisicionObservadores {
	CORR_EMPRESA: number;
	CORR_REQUISICION_PERSONAL: number;
	CORR_REQUISICION_OBSERVADORES: number;
	LOGIN_SISTEMA: string;
	TIPO_OBSERVADOR: string;
	FECHA_ASIGNACION: Date;
	ACTIVO: boolean;
	USUARIO_CREA: string;
	ESTACION_CREA: string;
	FECHA_CREA: Date;
	USUARIO_ACTU: string;
	ESTACION_ACTU: string;
	FECHA_ACTU: Date;
}